+++ lib/src/runner/loader.dart
@@ -127,14 +127,15 @@ class Loader {
     var controller = new StreamController();
     Future.forEach(_platforms, (platform) {
       if (!metadata.testOn.evaluate(platform, os: currentOS)) {
-        return new Future.value();
+        return;
       }
 
       // Don't load a skipped suite.
       if (metadata.skip) {
-        return new Future.value(new Suite([
+        controller.add(new Suite([
           new LocalTest(path, metadata, () {})
         ], path: path, platform: platform.name, metadata: metadata));
+        return;
       }
 
       return new Future.sync(() {